+++ tests/testthat/test-utilities-object-functions.R
@@ -1,13 +1,14 @@
 ## context(".calculateCensoredContribution")
 
-obsVsPredDf <- readr::read_csv(getTestDataFilePath("Aciclovir_obsVsPredDf.csv"))
+obsVsPredDf <- readr::read_csv(getTestDataFilePath("Aciclovir_obsVsPredDf.csv"),
+                               show_col_types = FALSE)
 
 obsDf <- obsVsPredDf[obsVsPredDf$dataType == "observed", ]
 predDf <- obsVsPredDf[obsVsPredDf$dataType == "simulated", ]
 
 test_that("`.calculateCensoredContribution` correctly calculates result with linear scaling", {
   obsDf$lloq <- 2.5
-  result <- ospsuite.parameteridentification:::.calculateCensoredContribution(
+  result <- .calculateCensoredContribution(
     observed = obsDf, simulated = predDf,
     scaling = "lin", linScaleCV = 0.25
   )
@@ -16,10 +17,10 @@ test_that("`.calculateCensoredContribution` correctly calculates result with lin
 
 test_that("`.calculateCensoredContribution` correctly calculates result with logarithmic scaling", {
   obsVsPredDf$lloq <- 2.5
-  obsVsPredDfLog <- ospsuite.parameteridentification:::.applyLogTransformation(obsVsPredDf)
+  obsVsPredDfLog <- .applyLogTransformation(obsVsPredDf)
   obsDfLog <- obsVsPredDfLog[obsVsPredDfLog$dataType == "observed", ]
   predDfLog <- obsVsPredDfLog[obsVsPredDfLog$dataType == "simulated", ]
-  result <- ospsuite.parameteridentification:::.calculateCensoredContribution(
+  result <- .calculateCensoredContribution(
     observed = obsDfLog, simulated = predDfLog,
     scaling = "log", logScaleSD = 0.086
   )
@@ -30,7 +31,7 @@ test_that("`.calculateCensoredContribution` can handle a minimal dataset with a 
   obsDf$lloq <- 2.5
   obsDfSingle <- obsDf[1, , drop = FALSE]
   predDfSingle <- predDf[1, , drop = FALSE]
-  result <- ospsuite.parameteridentification:::.calculateCensoredContribution(
+  result <- .calculateCensoredContribution(
     observed = obsDfSingle, simulated = predDfSingle,
     scaling = "lin", linScaleCV = 0.2
   )
@@ -39,14 +40,14 @@ test_that("`.calculateCensoredContribution` can handle a minimal dataset with a 
 
 test_that("`.calculateCensoredContribution` throws an error when LLOQ values are missing in the observed data", {
   expect_error(
-    result <- ospsuite.parameteridentification:::.calculateCensoredContribution(
+    result <- .calculateCensoredContribution(
       observed = obsDf, simulated = predDf,
       scaling = "lin", linScaleCV = 0.2
     )
   )
   obsDf$lloq <- NULL
   expect_error(
-    result <- ospsuite.parameteridentification:::.calculateCensoredContribution(
+    result <- .calculateCensoredContribution(
       observed = obsDf, simulated = predDf,
       scaling = "lin", linScaleCV = 0.2
     )
@@ -56,19 +57,19 @@ test_that("`.calculateCensoredContribution` throws an error when LLOQ values are
 test_that("`.calculateCensoredContribution` throws errors on invalid options", {
   obsDf$lloq <- 2.5
   expect_error(
-    result <- ospsuite.parameteridentification:::.calculateCensoredContribution(
+    result <- .calculateCensoredContribution(
       observed = obsDf, simulated = predDf,
       scaling = "invalidOption", linScaleCV = 0.2
     )
   )
   expect_error(
-    result <- ospsuite.parameteridentification:::.calculateCensoredContribution(
+    result <- .calculateCensoredContribution(
       observed = obsDf, simulated = predDf,
       scaling = "lin", logScaleSD = 0.086
     )
   )
   expect_error(
-    result <- ospsuite.parameteridentification:::.calculateCensoredContribution(
+    result <- .calculateCensoredContribution(
       observed = obsDf, simulated = predDf,
       scaling = "log", linScaleCV = 0.2
     )
@@ -77,10 +78,8 @@ test_that("`.calculateCensoredContribution` throws errors on invalid options", {
 
 ## context(".applyLogTransformation")
 
-obsVsPredDf <- readr::read_csv(getTestDataFilePath("Aciclovir_obsVsPredDf.csv"))
-
 test_that("`.applyLogTransformation` correctly log-transforms yValues and lloq", {
-  obsVsPredDfLog <- ospsuite.parameteridentification:::.applyLogTransformation(obsVsPredDf)
+  obsVsPredDfLog <- .applyLogTransformation(obsVsPredDf)
   expect_snapshot_value(obsVsPredDfLog, style = "serialize")
 })
 
@@ -90,7 +89,7 @@ test_that("`.calculateHuberWeights` calculates correct weights for standard resi
   residuals <- c(-2, -1, 0, 1, 2)
   expected_weights <- c(1, 1, 1, 1, 1)
   expect_equal(
-    ospsuite.parameteridentification:::.calculateHuberWeights(residuals),
+    .calculateHuberWeights(residuals),
     expected_weights,
     tolerance = 0.01
   )
@@ -98,7 +97,7 @@ test_that("`.calculateHuberWeights` calculates correct weights for standard resi
 
 test_that("`.calculateHuberWeights` returns empty vector for empty residuals", {
   residuals <- numeric(0)
-  expect_equal(ospsuite.parameteridentification:::.calculateHuberWeights(residuals), logical(0))
+  expect_equal(.calculateHuberWeights(residuals), logical(0))
 })
 
 ## context(.calculateBisquareWeights)
@@ -107,7 +106,7 @@ test_that("`.calculateBisquareWeights` calculates correct weights for standard r
   residuals <- c(-2, -1, 0, 1, 2)
   expected_weights <- c(0.841, 0.959, 1, 0.959, 0.841)
   expect_equal(
-    ospsuite.parameteridentification:::.calculateBisquareWeights(residuals),
+    .calculateBisquareWeights(residuals),
     expected_weights,
     tolerance = 0.01
   )
@@ -115,12 +114,10 @@ test_that("`.calculateBisquareWeights` calculates correct weights for standard r
 
 test_that("`.calculateBisquareWeights` returns empty vector for empty residuals", {
   residuals <- numeric(0)
-  expect_equal(ospsuite.parameteridentification:::.calculateHuberWeights(residuals), logical(0))
+  expect_equal(.calculateHuberWeights(residuals), logical(0))
 })
 
 ## context("calculateCostMetrics")
-
-obsVsPredDf <- readr::read_csv(getTestDataFilePath("Aciclovir_obsVsPredDf.csv"))
 
 test_that("'calculateCostMetrics' returns expected cost metrics for valid input data and default parameters", {
   result <- calculateCostMetrics(obsVsPredDf)